Refactor submodule related code
