+++ services/gitdiff/submodule.go
@@ -15,7 +15,7 @@ import (
 
 type SubmoduleDiffInfo struct {
 	SubmoduleName string
-	SubmoduleFile *git.CommitSubmoduleFile // it might be nil if the submodule is not found or unable to parse
+	SubmoduleFile *git.SubmoduleFile // it might be nil if the submodule is not found or unable to parse
 	NewRefID      string
 	PreviousRefID string
 }
@@ -37,7 +37,7 @@ func (si *SubmoduleDiffInfo) PopulateURL(repoLink string, diffFile *DiffFile, le
 		return // ignore the error, do not cause 500 errors for end users
 	}
 	if submodule != nil {
-		si.SubmoduleFile = git.NewCommitSubmoduleFile(repoLink, submoduleFullPath, submodule.URL, submoduleCommit.ID.String())
+		si.SubmoduleFile = git.NewSubmoduleFile(repoLink, submoduleFullPath, submodule.URL, submoduleCommit.ID.String())
 	}
 }
